refactor: move GradleInitScript to cli-common
