+++ devtools/cli/src/main/java/io/quarkus/cli/BuildToolDelegatingCommand.java
@@ -15,8 +15,8 @@ import io.quarkus.cli.common.HelpOption;
 import io.quarkus.cli.common.OutputOptionMixin;
 import io.quarkus.cli.common.PropertiesOptions;
 import io.quarkus.cli.common.RunModeOption;
+import io.quarkus.cli.common.gradle.GradleInitScript;
 import io.quarkus.cli.registry.ToggleRegistryClientMixin;
-import io.quarkus.cli.utils.GradleInitScript;
 import io.quarkus.devtools.project.BuildTool;
 import picocli.CommandLine;
 import picocli.CommandLine.ExitCode;
